perf: fix cross-platform GPU monitoring
Keep XPU-SMI tests on supported Unix targets and align multi-GPU samples by polling second.
